nrf_802154: rev dd2f456a2f55e6d0445dc4b1e9b30efac56b8bd4
This commit updates revision of the nrf_802154 component.

+++ nrf_802154/driver/src/mac_features/nrf_802154_tx_timestamp_provider.h
@@ -40,7 +40,7 @@
 
 #include "nrf_802154_types_internal.h"
 
-#define NRF_802154_TX_TIMESTAMP_PROVIDER_TIMESTAMP_SIZE (sizeof(uint64_t))
+#define NRF_802154_TX_TIMESTAMP_PROVIDER_TIMESTAMP_SIZE (sizeof(uint32_t))
 
 /**
  * @brief TX setup hook for the tx timestamp provider module.
